feat: splitted pagination comp code
